requirements.txt
fastapi
uvicorn[standard]
gunicorn
SQLAlchemy
python-dotenv
passlib[bcrypt]
psycopg2-binary
pydantic
PyYAML
starlette
pytest
pytest-cov
anyio
websockets
watchfiles
typing-extensions
python-jose
email-validator
